pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["carterwu1120 <carterwu1120@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
speechrecognition = "^3.12.0"
gradio = "^5.8.0"
sounddevice = "^0.5.1"
scipy = "^1.14.1"
pyaudio = "^0.2.14"
poetry = "^1.8.5"
setuptools = "^75.6.0"
numpy = "2.0"
ollama = "^0.4.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
